Move duplicate interpolate files into bruno/js module
