+++ packages/bruno-cli/src/runner/run-single-request.js
@@ -6,8 +6,8 @@ const fs = require('fs');
 const { forOwn, isUndefined, isNull, each, extend, get, compact } = require('lodash');
 const FormData = require('form-data');
 const prepareRequest = require('./prepare-request');
-const interpolateVars = require('./interpolate-vars');
-const { interpolateString } = require('./interpolate-string');
+const interpolateVars = require('@usebruno/js/src/interpolate-vars');
+const { interpolateString } = require('@usebruno/js/src/interpolate-string');
 const { ScriptRuntime, TestRuntime, VarsRuntime, AssertRuntime } = require('@usebruno/js');
 const { stripExtension } = require('../utils/filesystem');
 const { getOptions } = require('../utils/bru');
