+++ packages/bruno-electron/src/ipc/network/index.js
@@ -15,8 +15,8 @@ const prepareRequest = require('./prepare-request');
 const prepareGqlIntrospectionRequest = require('./prepare-gql-introspection-request');
 const { cancelTokens, saveCancelToken, deleteCancelToken } = require('../../utils/cancel-token');
 const { uuid } = require('../../utils/common');
-const interpolateVars = require('./interpolate-vars');
-const { interpolateString } = require('./interpolate-string');
+const interpolateVars = require('@usebruno/js/src/interpolate-vars');
+const { interpolateString } = require('@usebruno/js/src/interpolate-string');
 const { sortFolder, getAllRequestsInFolderRecursively } = require('./helper');
 const { preferencesUtil } = require('../../store/preferences');
 const { getProcessEnvVars } = require('../../store/process-env');
