+++ packages/bruno-js/src/index.js
@@ -2,10 +2,14 @@ const ScriptRuntime = require('./runtime/script-runtime');
 const TestRuntime = require('./runtime/test-runtime');
 const VarsRuntime = require('./runtime/vars-runtime');
 const AssertRuntime = require('./runtime/assert-runtime');
+const { InterpolateUrl } = require('./interpolate-string');
+const { InterpolateVars } = require('./interpolate-vars');
 
 module.exports = {
   ScriptRuntime,
   TestRuntime,
   VarsRuntime,
-  AssertRuntime
+  AssertRuntime,
+  InterpolateUrl,
+  InterpolateVars
 };
